+++ src/components/TextEditor/style.tsx
@@ -1,10 +1,22 @@
 import styled from 'styled-components';
+import dynamic from 'next/dynamic';
+import React from 'react';
 
-export const StyledQuillEditor = styled.div`
-	width: 100%;
-	height: 100%;
+const QuillWrapper = dynamic(() => import('react-quill'), {
+	ssr: false,
+	loading: () => <p>Loading ...</p>,
+});
+
+export const StyledQuillWrapper = styled(QuillWrapper)<{ minHeight?: string, borderRadius?: string }>`
+	.ql-toolbar {
+		border-radius: ${(props) => props.borderRadius || '0'} ${(props) => props.borderRadius || '0'} 0 0;
+    border: 1px solid #c9c9c9;
+  }
 	.ql-editor {
-		width: 100%;
-		height: 100%;
+		min-height: ${(props) => props.minHeight || 'auto'};
 	}
+	.ql-container {
+		border-radius: 0 0 ${(props) => props.borderRadius || '0'} ${(props) => props.borderRadius || '0'};
+    border: 1px solid #c9c9c9;
+  }
 `;
